desgin4green/requirements.txt
pandas==1.4.4
Pillow==9.3.0
streamlit==1.14.0
~treamlit==1.13.0
